chore(quality): PHPStan level 8, Psalm errorLevel 4, Codecov, Rector migration
Quality bar raised across the board on the v4 namespace.

Static analysis:
- PHPStan bumped from level 6 → level 8 on src/v4 + tests/v4. Required
  fixes: typed-property nullability initialisation in RetryClient (move
  $response/$lastResponse out of the try block), explicit float casts in
  ExponentialBackoff / SystemSleeper (level-8 strict binary operand mode),
  and tightened @param Endpoint<mixed> generics across AsyncTransport.
- Psalm added as a second-opinion analyser (vimeo/psalm:^6 — PHP 8.4
  support landed in 6.x). Configured at errorLevel 4, scoped to src/v4 +
  tests/v4. Psalm autofix added #[\Override] attributes on every PSR-
  override / parent-override method (115+ files touched).
- Real bugs Psalm caught: a SimpleXMLElement null-coalesce that can never
  be null (BringApiException XML parser), a redundant getResponse() !==
  null check (Guzzle 7 returns non-nullable), and a never-initialised
  variable on the give-up branch of RetryClient.

CI:
- New "Psalm" job alongside the existing PHPStan job, running with
  --output-format=github so issues annotate the diff directly.
- "test" job now runs PHPUnit with pcov coverage on the PHP 8.3 / highest
  matrix entry and uploads to Codecov. codecov.yml pins the v4 namespace
  as the watched target with an 80% threshold; legacy v3 directories are
  ignored so the project rating tracks the modern surface.
- composer.json gains "test-coverage", "psalm", and "qa" (umbrella
  cs+phpstan+psalm+test) scripts.

Migration tooling:
- New rector-v4-migration.php ships a RenameClassRector mapping that
  downstream applications can run with:
    vendor/bin/rector process src \
      --config vendor/crakter/bringapi/rector-v4-migration.php
  to mechanically rewrite v3 class references to their v4 facade
  equivalents. The constructor-shape-changing parts (fluent setter chains
  to typed DTOs) still need manual review — but the namespace migration,
  which is most of the touch-points in a typical adoption, is automatic.

Documentation:
- README "Development" section refreshed to use the composer-script
  aliases; UPGRADE-4.0.md gains a "Automated migration with Rector"
  section listing exactly what the ruleset does and does not handle.

172 → 183 PHPUnit tests still pass; PHPStan level-8 clean; Psalm clean.

+++ src/v4/Retry/ExponentialBackoff.php
@@ -27,12 +27,13 @@ final class ExponentialBackoff implements BackoffStrategy
         $this->rng = $rng ?? static fn (): float => mt_rand() / mt_getrandmax();
     }
 
+    #[\Override]
     public function delaySeconds(int $attempt): float
     {
         if ($attempt < 1) {
             return 0.0;
         }
-        $exp = min($this->maxDelaySeconds, $this->baseDelaySeconds * (2 ** ($attempt - 1)));
+        $exp = min($this->maxDelaySeconds, $this->baseDelaySeconds * (float) (2 ** ($attempt - 1)));
 
         return ($this->rng)() * $exp;
     }
